Introduce `BaseDataManager` as a new base class.
...because those methods didn't really belong on either `BaseComplexMember`
or `BaseControl`, and the `null`-returning behavior seemed wonky and possibly
even surprising.
